update new version, clean code, add role, permission

+++ app/services/user_service.py
@@ -1,11 +1,16 @@
 from db import db
 from models.user_model import UserModel
+from models.role_model import RoleModel
 from models.blocklist_model import BlocklistModel
 from datetime import datetime
 from passlib.hash import pbkdf2_sha256
 from sqlalchemy import asc
-from flask_jwt_extended import create_access_token, create_refresh_token
+from flask import current_app
+from flask_principal import identity_changed, Identity
+from flask_jwt_extended import create_access_token, create_refresh_token, get_jwt_identity, get_jwt
 from flask_smorest import abort
+from config import DEFAULT_ROLE
+from services import user_role_service
 
 def get_all_user():
     results = UserModel.query.order_by(asc(UserModel.id)).all()
@@ -16,21 +21,25 @@ def get_user(user_id):
     return results
 
 def update_user(user_data, user_id):
-    result = UserModel.query.filter_by(id=user_id).first()
-    if not result:
+    user = UserModel.query.filter_by(id=user_id).first()
+    if not user:
         abort(400, message="User doesn't exist, cannot update!")
 
     try:
         if user_data['username']:
-            result.username = user_data['username']
+            user.username = user_data['username']
             
         if user_data['password']:
             # Hash password
             password = pbkdf2_sha256.hash(user_data['password'])
-            result.password = password
-            
-        if user_data['time_created']:
-            result.time_created = user_data['time_created']
+            user.password = password
+        
+        # Update roles  
+        user.roles = []
+        
+        for role_id in user_data['roles']:
+            role = RoleModel.query.filter_by(id=role_id).first()
+            user.roles.append(role)
             
         db.session.commit()
     except:
@@ -39,7 +48,33 @@ def update_user(user_data, user_id):
 
     return {"message": "Update successfully!"}
 
+def update_block_user(user_data, user_id):   
+    if user_id == 1:
+        abort(401, message="Can not block Super Admin!")
+    
+    user = UserModel.query.filter_by(id=user_id).first()
+    
+    if not user:
+        abort(400, message="User doesn't exist, cannot update!")
+
+    try:         
+        # Update status block
+        user.block = user_data['block']
+        
+        db.session.add(user)
+        db.session.commit()
+    except:
+        db.session.rollback()
+        abort(400, message="Can not update status block User!")
+
+    return {"message": "Block successfully!"}
+
 def delete_user(id):
+    # Only admin can delete user
+    jwt = get_jwt()
+    if not jwt.get("is_admin"):
+        abort(401, message="Admin privilege requierd.")
+    
     result = UserModel.query.filter_by(id=id).delete()
     if not result:
         abort(400, message="User doesn't exist, cannot delete!")
@@ -59,6 +94,10 @@ def login_user(user_data):
         # Create refresh_token
         refresh_token = create_refresh_token(identity=user.id)
         
+        # Set the user identity to the role
+        identity = Identity(user.id)
+        identity_changed.send(current_app._get_current_object(), identity=identity)
+        
         return {"access_token": access_token, "refresh_token": refresh_token}
 
     abort(401, message="Invalid credentials.")
@@ -66,7 +105,6 @@ def login_user(user_data):
 def register_user(user_data):
     username = user_data['username']
     password = user_data['password']
-    time_created = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
     
     # Check user name exist
     user = UserModel.query.filter(UserModel.username == user_data['username']).first()
@@ -76,17 +114,43 @@ def register_user(user_data):
     # Hash password
     password = pbkdf2_sha256.hash(password)
     
-    new_row = UserModel(username = username, password = password, time_created = time_created)
-    
+    new_user = UserModel(username = username, password = password)
+
     try:
-        db.session.add(new_row)
+        db.session.add(new_user)
         db.session.commit()
+    
     except:
         db.session.rollback()
         abort(400, message="Can not register!")
+        
+    try:
+        # Add default role for user
+        user_role_service.link_roles_to_user(user_id=new_user.id, role_id=DEFAULT_ROLE)
+        
+    except:
+        abort(400, message="Can not register! - Can not add role default.")
     
-    return {"message": "Add successfully!"}
+    return {"message": "Register successfully!"}
 
+def refresh_token():
+    # Get id current user
+    current_user_id = get_jwt_identity()
+    
+    # Create access_token
+    access_token = create_access_token(identity=current_user_id, fresh=True)
+    
+    # Create refresh_token
+    refresh_token = create_refresh_token(identity=current_user_id)
+    
+    # Block previous access_token
+    jti = get_jwt()['jti']
+    
+    # Block access token
+    add_jti_blocklist(jti)
+    
+    return {"access_token": access_token, "refresh_token": refresh_token}
+    
 def add_jti_blocklist(jti):
     # Add to blockist when remove jti
     new_row = BlocklistModel(jti_blocklist = str(jti))
